Navigate to alternate file with alt
Requres https://github.com/uptech/alt

vim-rails provides this functionality and I use it all the time.

I've had the generalized setup with alt in my .vimrc for a while, but I
don't know if I use it and don't realize or if I don't need such
functionality outside of Rails.